+++ hawkbit-repository/hawkbit-repository-jpa/src/main/java/org/eclipse/hawkbit/repository/jpa/utils/QuotaHelper.java
@@ -15,12 +15,10 @@ import javax.validation.constraints.NotNull;
 import org.eclipse.hawkbit.repository.exception.QuotaExceededException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.validation.annotation.Validated;
 
 /**
  * Helper class to check assignment quotas.
  */
-@Validated
 public final class QuotaHelper {
 
     /**
